made id to be integer and define a enum into entity

+++ PosProject/src/main/java/com/pos/posproject/entity/Payment.java
@@ -1,3 +1,8 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
 package com.pos.posproject.entity;
 
 import com.pos.posproject.enums.PaymentMethods;
@@ -18,20 +23,20 @@ public class Payment implements Serializable {
     private static final long serialVersionUID = 1L;
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private Integer id;
     
     private Double amount;
     private Date date;
     private PaymentMethods method;
-    
-    public Long getId() {
+
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
-    
+
     public Double getAmount() {
         return amount;
     }
